requirements.txt
flask
python-telegram-bot==13.15
requests
opencv-python
numpy
pillow
face-recognition
dlib
python-dotenv
